+++ tests/automapping/test_automapping.cpp
@@ -22,6 +22,7 @@ void test_AutoMapping::autoMap_data()
 {
     QTest::addColumn<QString>("directory");
 
+    QTest::newRow("ignore-empty-region") << QStringLiteral("ignore-empty-region");
     QTest::newRow("ignore-flip") << QStringLiteral("ignore-flip");
     QTest::newRow("infinite-target-map") << QStringLiteral("infinite-target-map");
     QTest::newRow("inputnot") << QStringLiteral("inputnot");